+++ demos/storybook/storybook/stories/channel-value.tsx
@@ -7,8 +7,7 @@ import MatIcon from 'react-native-vector-icons/MaterialIcons';
 import * as Colors from '@pxblue/colors';
 
 const notes = {
-    notes:
-        'Any React Element may be passed in as `icon`; if using an svg, its color and size are not controlled by `ChannelValue`',
+    notes: 'Any React Element may be passed in as `icon`; if using an svg, its color and size are not controlled by `ChannelValue`',
 };
 
 const WrappedTrending = wrapIcon({ IconClass: MatIcon, name: 'trending-up' });