+++ demos/storybook/storybook/stories/empty-state.tsx
@@ -41,8 +41,7 @@ storiesOf('EmptyState', module)
             source={
                 boolean('Show Background', true)
                     ? {
-                          uri:
-                              'https://images.unsplash.com/photo-1500382017468-9049fed747ef?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&w=1000&q=80',
+                          uri: 'https://images.unsplash.com/photo-1500382017468-9049fed747ef?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&w=1000&q=80',
                       }
                     : {}
             }